requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
scipy==1.7.2
pandas==1.3.5